components/pages/Trip.js: moved, updated imports
