+++ src/components/pages/Trip.js
